Rename traits::drop -> traits::misc.
